chore: move activity piece to community
