+++ packages/pieces/community/activity/project.json
@@ -1,0 +1,33 @@
+{
+  "name": "pieces-activity",
+  "$schema": "../../../../node_modules/nx/schemas/project-schema.json",
+  "sourceRoot": "packages/pieces/community/activity/src",
+  "projectType": "library",
+  "targets": {
+    "build": {
+      "executor": "@nx/js:tsc",
+      "outputs": ["{options.outputPath}"],
+      "options": {
+        "outputPath": "dist/packages/pieces/community/activity",
+        "tsConfig": "packages/pieces/community/activity/tsconfig.lib.json",
+        "packageJson": "packages/pieces/community/activity/package.json",
+        "main": "packages/pieces/community/activity/src/index.ts",
+        "assets": ["packages/pieces/community/activity/*.md"],
+        "buildableProjectDepsInPackageJsonType": "dependencies",
+        "updateBuildableProjectDepsInPackageJson": true
+      }
+    },
+    "publish": {
+      "command": "node tools/scripts/publish.mjs pieces-activity {args.ver} {args.tag}",
+      "dependsOn": ["build"]
+    },
+    "lint": {
+      "executor": "@nx/eslint:lint",
+      "outputs": ["{options.outputFile}"],
+      "options": {
+        "lintFilePatterns": ["packages/pieces/community/activity/**/*.ts"]
+      }
+    }
+  },
+  "tags": []
+}
