+++ packages/pieces/community/activity/tsconfig.lib.json
